requirements.txt
# GeoGuessr Country Classification - Requirements
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0

# Computer vision
opencv-python>=4.8.0
scikit-image>=0.21.0

# Machine learning
scikit-learn>=1.3.0
numpy>=1.24.0

# Data handling
pandas>=2.0.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

# Optional: for Grad-CAM visualization
grad-cam>=1.4.8

